'Decently' thought out controls for the shooter
The following controls have been implemented:

'A': Tracks values(constantly) from a sensor and adjusts a speed based on a [not yet written] function / Stops motors

'B': Sets the shooter to the 'default' speed / Locks in the speed set by 'A'(this is required for making adjustments to the speed manually) / Stops motors

'UP': Increases the speed of the motors while held.

'DOWN': Decreases the speed of the motors while held.